Do not duplicate `TransactionContext`
PR-710
